% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acgh}
\alias{acgh}
\title{Microarray data.}
\format{A dataframe with 2215 rows and 41 columns. Each column contains data for one individual.}
\usage{
data(acgh)
}
\description{
Microarray data for 41 individuals with a bladder tumour. The data consists of a log-likelihood-ratio statistic testing for the presence of a copy 
number variation evaluated along each individual's genome. The data was obtained from the \pkg{ecp} package (James and Mattesson, 2014) and standardised using
robust measures of mean and variance.
}
\references{
\insertRef{ecp-package}{anomaly}
}
\keyword{datasets}
