% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/anomaly1.R
\name{anomalyComp}
\alias{anomalyComp}
\title{anomalyComp}
\usage{
anomalyComp(MTarget, MComp, breaks = 4, type = "Core", sigma = 2)
}
\arguments{
\item{breaks}{The maximum number of breaks}

\item{dates}{Vector of numeric time}

\item{values}{Target vector of values}

\item{tall}{table of comparators}

\item{fitted}{Do you want fitted values returned}
}
\value{
breaks data.frame of breaks and location

data table of values, and fitted values
}
\description{
anomaly tests against a data.table of comparators
by fitting breaks to residuals and combining
outputs data.table of fitted breaks.
}

