% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bard.R
\name{bard}
\alias{bard}
\title{Detection of multivariate anomalous segments using BARD.}
\usage{
bard(x, p_N, p_A, k_N, k_A, pi_N, alpha, paffected, lower, upper, h,
  transform = robustscale)
}
\arguments{
\item{x}{- An $n$ by $p$ real matrix representing n observations of p variates. Each variate is scaled by BARD using the median and the median absolute deviation. This
can be changed using the \code{transform} parameter.}

\item{p_N}{- Probability of success in each trial for the Negative Binomial distribution for the length of normal segments.}

\item{p_A}{- Probability of success in each trial for the Negative Binomial distribution for the length of abnormal segments.}

\item{k_N}{- Dispersion parameter for the Negative Binomial distribution for the length of normal segments.}

\item{k_A}{- Dispersion parameter for the Negative Binomial distribution for the length of abnormal segments.}

\item{pi_N}{- Probability that an abnormal segment is followed by a normal segment.}

\item{alpha}{- Threshold used to control the resampling in the approximation of the posterior distribution at each time step.}

\item{paffected}{- Proportion of the series believed to be affected by an abnormal segment.}

\item{lower}{- The lower limit of the prior uniform distribution for \eqn{mu}.}

\item{upper}{- The upper limit of the prior uniform distribution for \eqn{mu}.}

\item{h}{- The step size in the numerical integration used to find the marginal likelihood. The quadrature points are located from lower to upper in steps of h.}

\item{transform}{- A function used to transform the data prior to analysis. The default value is to scale the data using the median and the median absolute deviation.}
}
\value{
An instance of the S4 object of type \code{.bard.class} containing the data \code{x}, procedure parameter values, and the results.
}
\description{
Implements the BARD (Bayesian Abnormal Region Detector) procedure of Bardwell and Fearnhead (2017). BARD is a fully Bayesian inference procedure which is able to give
measures of uncertainty about the number and location of abnormal regions. It uses flexible prior distributions on the lengths of normal and abnormal regions as well as
a prior distribution on the mean of the affected variates.
}
\examples{

library(anomaly)
set.seed(0)
sim.data<-simulate(n=500,p=200,mu=2,locations=c(100,200,300),
                   duration=6,proportions=c(0.04,0.06,0.08))
# parameters
p_N<-0.05
p_A<-0.5
k_N<-10
k_A<-10
pi_N<-0.9
alpha<-0.0001
paffected<-10/200
lower<-0.5
upper<-1.5
h<-0.25
# run bard
res<-bard(sim.data,p_N,p_A,k_N,k_A,pi_N,alpha,paffected,lower,upper,h)
sampler(res)

}
\references{
\insertRef{bardwell2017}{anomaly}
}
