\name{Kepler10965588}
\docType{data}
\alias{Kepler10965588}
\title{
Lightcurve for the star Kepler 10965588. 
}
\description{
Kepler 10965588. orbited by one known planet with orbital period of 62.89 days.

This research has made use of the NASA Exoplanet Archive, which is operated by the California Institute of Technology, under contract with the National Aeronautics and Space Administration under the Exoplanet Exploration Program.

}
\usage{
data(Lightcurves)
}
\format{
A data frame with two columns called "Brightness" and "Day", containing measurements of a star's brightness and the measurement's timestamp respectively. 
}
\keyword{datasets}