\name{Kepler1871056}
\docType{data}
\alias{Kepler1871056}
\title{
Lightcurve for the star Kepler 1871056. 
}
\description{
Kepler 1871056. orbited by two known planets with orbital periods of 40.8 and 140.1 days.

This research has made use of the NASA Exoplanet Archive, which is operated by the California Institute of Technology, under contract with the National Aeronautics and Space Administration under the Exoplanet Exploration Program.

}
\usage{
data(Lightcurves)
}
\format{
A data frame with two columns called "Brightness" and "Day", containing measurements of a star's brightness and the measurement's timestamp respectively. 
}
\keyword{datasets}