\name{Kepler3102384}
\docType{data}
\alias{Kepler3102384}
\title{
Lightcurve for the star Kepler 3102384. 
}
\description{
Kepler 3102384. orbited by two known planets with orbital periods of 10.57 and 523.9 days

This research has made use of the NASA Exoplanet Archive, which is operated by the California Institute of Technology, under contract with the National Aeronautics and Space Administration under the Exoplanet Exploration Program.

}
\usage{
data(Lightcurves)
}
\format{
A data frame with two columns called "Brightness" and "Day", containing measurements of a star's brightness and the measurement's timestamp respectively. 
}
\keyword{datasets}