\name{Kepler3231341}
\docType{data}
\alias{Kepler3231341}
\title{
Lightcurve for the star Kepler 3231341. 
}
\description{
Kepler 3231341. orbited by four known planets with orbital periods of 4.24, 8.15, 12.33, and 19.00 days

This research has made use of the NASA Exoplanet Archive, which is operated by the California Institute of Technology, under contract with the National Aeronautics and Space Administration under the Exoplanet Exploration Program.

}
\usage{
data(Lightcurves)
}
\format{
A data frame with two columns called "Brightness" and "Day", containing measurements of a star's brightness and the measurement's timestamp respectively. 
}
\keyword{datasets}