\name{Kepler3447722}
\docType{data}
\alias{Kepler3447722}
\title{
Lightcurve for the star Kepler 3447722. 
}
\description{
Kepler 3447722. orbited by three known planets with orbital periods of 10.30, 16.09, and 35.68 days

This research has made use of the NASA Exoplanet Archive, which is operated by the California Institute of Technology, under contract with the National Aeronautics and Space Administration under the Exoplanet Exploration Program.

}
\usage{
data(Lightcurves)
}
\format{
A data frame with two columns called "Brightness" and "Day", containing measurements of a star's brightness and the measurement's timestamp respectively. 
}
\keyword{datasets}