\name{Kepler4139816}
\docType{data}
\alias{Kepler4139816}
\title{
Lightcurve for the star Kepler 4139816. 
}
\description{
Kepler 4139816. orbited by four known planets with orbital periods of 3.34, 7.82, 20.06, and 46.18 days

This research has made use of the NASA Exoplanet Archive, which is operated by the California Institute of Technology, under contract with the National Aeronautics and Space Administration under the Exoplanet Exploration Program.

}
\usage{
data(Lightcurves)
}
\format{
A data frame with two columns called "Brightness" and "Day", containing measurements of a star's brightness and the measurement's timestamp respectively. 
}
\keyword{datasets}