% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{machinetemp}
\alias{machinetemp}
\title{Machine temperature data.}
\format{
A dataframe with 22695 rows and 2 columns. The first column contains the date and time of the temperature measurement. The second column contains the
machine temperature.
}
\usage{
data(machinetemp)
}
\description{
Temperature sensor data of an internal component of a large, industrial machine. The data contains three known anomalies.
The first anomaly is a planned shutdown of the machine. The second anomaly is difficult to detect and directly led to the third anomaly, a catastrophic failure of the machine.
The data consists of 22695 observations of machine temperature recorded at 5 minute intervals along with the date and time of the measurement. The data was obtained from the Numenta
Anomaly Benchmark \insertCite{AHMAD2017134}{anomaly}, which can be found at https://github.com/numenta/NAB.
}
\references{
\insertRef{AHMAD2017134}{anomaly}
}
\keyword{datasets}
