% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/anomaly1.R
\name{model.ss}
\alias{model.ss}
\title{model.ss}
\usage{
model.ss(D, max.breaks = 4, sig, plot = FALSE)
}
\arguments{
\item{sig}{is the critical value for accepting}

\item{plot}{for plotting intermediate results}

\item{y}{timeseries or vector of data}

\item{d}{The maximum number of breaks is 3^d}
}
\value{
cuts the data.table of cuts
}
\description{
model.ss takes the data as a data.frame and returns cuts
should this be called something else?
}

