% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pass.R
\name{pass}
\alias{pass}
\title{Proportion Adaptive Segment Selection (PASS).}
\usage{
pass(Xi, Lmax, alpha, lambda, Lmin = 0)
}
\arguments{
\item{Xi}{An n x N real matrix representing n observations of N variates. NOTE(should this be normalised in some way ?)}

\item{Lmax}{A positive integer > 0 indicating the maximum segment. (Is there a sensible/informed way of choosing a default ?)}

\item{alpha}{A positive integer > 1. This value is used to stabilise the PASS procedure and control the familywise error for finite samples. (How to choose - what default ?).}

\item{lambda}{A positive real threshold value used to control the familywise Type 1 error to the desired level.}

\item{Lmin}{A positive integer < Lmax indicating the minimum segment length. Default value is \code{Lmax=1}.}
}
\value{
An S4 object of type \code{.pass.class} containing the data \code{Xi}, procedure parameter values and the results.
}
\description{
Implements the Proportion Adaptive Segment Selection procedure of REF HERE to (what does it do ???)
}
\examples{
library(anomaly)
data(simulated)
res<-pass(sim.data,200,2,15.0)
summary(res)
plot(res,variate_names=FALSE)

}
\references{
\insertRef{10.1093/biomet/ass059}{anomaly}
}
