% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_series.R, R/capa.R, R/pass.class.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot.anomaly_series}
\alias{plot,anomaly_series-method}
\alias{plot,capa.class,ANY-method}
\alias{plot,capa.uv.class,ANY-method}
\alias{plot,capa.mv.class,ANY-method}
\alias{plot-pass.class}
\alias{plot,pass.class,ANY-method}
\title{Plots an anomaly_series object.}
\usage{
plot.anomaly_series(x, xlab = "", ylab = "", ...)

\S4method{plot}{capa.class,ANY}(x, subset, variate_names, tile_plot)

\S4method{plot}{capa.uv.class,ANY}(x, variate_name)

\S4method{plot}{capa.mv.class,ANY}(x, subset, variate_names, tile_plot)

\S4method{plot}{pass.class,ANY}(x, subset, variate_names, tile_plot)
}
\arguments{
\item{x}{An S4 class produced by \code{\link{capa}}, \code{\link{capa.uv}}, \code{\link{capa.mv}} or an S3 class produced by \code{\link{anomaly_series}}.}

\item{xlab}{Character string containing label for the x-axis.}

\item{ylab}{Character string containing label for the y-axis.}

\item{...}{Other parameters to be passed to plotting methods.}

\item{subset}{A numeric vector specifying a subset of the variates to be displayed. Default value is all of the variates present in the data.}

\item{variate_names}{Logical value indicating if variate names should be displayed on the plot. This is useful when a large number of variates are being displayed
as it makes the visualisation easier to interpret. Default value is TRUE.}

\item{tile_plot}{Logical value. If TRUE then a tile plot of the data is produced. The data displayed in the tile plot is normalised to values in [0,1] for each variate.
This type of plot is useful when the data contains are large number of variates. The default value is TRUE if the number of variates is greater than 20.}

\item{variate_name}{Logical value indicating if the variate name should be displayed. Default value is \code{variate.name=TRUE}.}
}
\value{
A ggplot object.
}
\description{
Plot method for anomaly_series objects as returned by the \code{\link{anomaly_series}} method.

Plot methods for S4 objects returned by \code{\link{capa}}, \code{\link{capa.uv}}, \code{\link{capa.mv}}, and \code{\link{pass}}. 

The plot can either be a line plot or a tile plot, the type produced depending on the options provided to the \code{plot} function and/or the dimensions of the
data associated with the S4 object.
}
\examples{
\donttest{
library(anomaly)
set.seed(2018)
# Generate data typically followig a normal distribution with mean 0 and variance 1. 
# Then introduce 3 anomaly windows and 4 point outliers.
x  = rnorm(5000)
x[401:500]   = rnorm(100,4,1)
x[1601:1800] = rnorm(200,0,0.01)
x[3201:3500] = rnorm(300,0,10)
x[c(1000,2000,3000,4000)] = rnorm(4,0,100)
inferred_anomalies = anomaly_series(x)
plot(inferred_anomalies)
}

}
\seealso{
\code{\link{capa}},\code{\link{capa.uv}},\code{\link{capa.mv}},\code{\link{pass}}.
}
