% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/anomaly1.R
\name{predict.ss}
\alias{predict.ss}
\title{predict.ss}
\usage{
\method{predict}{ss}(cuts, D, sig = 0.85, peak.ratio = 0)
}
\arguments{
\item{cuts}{the cutfile}

\item{sig}{is the critical value for accepting}

\item{data}{}
}
\value{
fitted values

breaks table

cuts cut table
}
\description{
predict.ss is fast breakpoints function
allowing for bid dtasets This combines
predict.ss and summary.ss
}

