% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_series.R
\name{print}
\alias{print}
\alias{print.anomaly_series}
\title{Print function for anomaly_series objects.}
\usage{
\method{print}{anomaly_series}(x, ...)
}
\arguments{
\item{x}{An anomaly_series object}

\item{...}{Other parameters to be passed to print methods.}
}
\description{
Prints the contents of an anomaly_series object as returned by the \code{\link{anomaly_series}} method.
}
