% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim.data}
\alias{sim.data}
\title{Simulated data.}
\format{A matrix with 500 rows and 200 columns.}
\usage{
data(simulated)
}
\description{
Simulated data for 200 variates with 500 observations containing three recurrent variants located at [100: 105], [200, 205], and [300:305]
with carrier proportions 0.04, 0.06, and 0.08 respectively. (add reference here)
}
\references{
\insertRef{10.1093/biomet/ass059}{anomaly}
}
\keyword{datasets}
