% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sos}
\alias{sos}
\title{Perform stochastic outlier selection on a distance matrix}
\usage{
sos(dist, perplexity)
}
\arguments{
\item{dist}{A distance matrix or a dist object as returned by \code{\link[stats]{dist}}}

\item{perplexity}{The size of the neighborhood to be considered}
}
\value{
A list with the following entries:
\describe{
 \item{outlierProbabilty}{A vector of outlier probabilities}
 \item{bindingProbability}{A matrix with the binding probabilities between all point}
}
}
\description{
This function calculates for each point in a distance matrix its probabilty
of being an outlier, given a certain perplexity. The perplexity parameter can
be seen as the size of the neighborhood taken into account when assessing the
outlierness of a given point, though contrary to k in KNNDD and LOF it does
not have to be an integer but can take any positive value (should be lower
than the number of points though). By setting a probability threshold the
output can be converted to a hard outlier classification.
}
\examples{
irisDist <- dist(iris[, 1:4])
irisSos <- sos(irisDist, 65)
irisMDS <- cmdscale(irisDist)
plot(irisMDS, pch=19, col=ifelse(irisSos$outlierProbabilty > 0.55, 'red', 'black'))
}

