% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/anomaly1.R
\name{summary.ss}
\alias{summary.ss}
\title{summary.ss}
\usage{
summary.ss(cuts, D)
}
\arguments{
\item{cuts}{the cutfile}

\item{data}{}
}
\value{
data with fitted values
}
\description{
summary.ss is fast breakpoints function
allowing for bid dtasets This combines
predict.ss and summary.ss
}

