% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/anomaly1.R
\name{sumseg}
\alias{sumseg}
\title{sumseg}
\usage{
sumseg(D)
}
\arguments{
\item{D}{timeseries or vector of data}

\item{max.breaks}{The maximum number of breaks is 3^d}

\item{sig}{is the critical value for accepting}

\item{plot}{for plotting intermediate results}
}
\value{
model fitted and the data.table of cuts
}
\description{
sumseg is fast breakpoints function
allowing for bid dtasets This combines
predict.ss and summary.ss
}
\examples{
values=rnorm(400)+c(rep(-0.5,400/2),rep(0.5,400/2))
o=sumseg(values)
plot(1:400,values)
lines(o$data$index,o$data$fitted,col=2,lwd=4)
print(o$breaks)
}

