% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_corrected.R
\name{ac_corrected}
\alias{ac_corrected}
\title{ac_corrected}
\usage{
ac_corrected(X)
}
\arguments{
\item{X}{A numeric matrix containing the data to be transformed.}
}
\value{
A numeric matrix containing the transformed data.
}
\description{
Transforms the data X to account for autocorrelation by centring and scaling using \eqn{X_{i}^{'} = \frac{X_{i}-\mu_{i}}{k_{i}\sigma_{i}}}. Here \eqn{\mu_{i}} and \eqn{\sigma_{i}} are robust estimates for
the mean and standard deviation of each variate (column) \eqn{X_{i}} of X calculated using the median and median absolute deviation and \eqn{k_{i} = \surd{\left( \frac{1+\phi_{i}}{1-\phi_{i}} \right)}} with
\eqn{\phi_{i}} a robust estimate for the autocorrelation.
}
\examples{
library(anomaly)
data(simulated)
# compare the medians of each variate and transformed variate
head(apply(sim.data,2,median))
head(apply(ac_corrected(sim.data),2,median))
# compare the variances of each variate and transformed variate
head(apply(sim.data,2,var))
head(apply(ac_corrected(sim.data),2,var))

}
