% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acgh}
\alias{acgh}
\title{Microarray data
Microarray data for 41 individuals with a bladder tumour. The data was obtained from the \pkg{ecp} package \insertRef{ecp-package}{anomaly} and standardised using
robust measures of mean and variance.}
\format{A dataframe with 2215 rows and 41 columns. Each column contains data for one individual.}
\usage{
data(acgh)
}
\description{
Microarray data
Microarray data for 41 individuals with a bladder tumour. The data was obtained from the \pkg{ecp} package \insertRef{ecp-package}{anomaly} and standardised using
robust measures of mean and variance.
}
\references{
\insertRef{ecp-package}{anomaly}
}
\keyword{datasets}
