% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pass.R
\name{pass}
\alias{pass}
\title{Proportion Adaptive Segment Selection (PASS).}
\usage{
pass(x, alpha = 2, lambda = 15, Lmax = 10, Lmin = 0,
  transform = robustscale)
}
\arguments{
\item{x}{An n x m real matrix representing n observations of m variates.}

\item{alpha}{A positive integer > 1. This value is used to stabilise the PASS procedure and control the familywise error for finite samples.}

\item{lambda}{A positive real value setting the threshold value for the familywise Type 1 error. The default value is 15.0.}

\item{Lmax}{A positive integer (\code{Lmax} > 1) corresponding to the maximum segment length. The default value is 10.}

\item{Lmin}{A positive integer (\code{Lmax} > \code{Lmin} > 1) corresponding to the minimum segment length. The default value is 0. .}

\item{transform}{A function used to transform the data prior to analysis. The default value is to scale the data using the median and the median absolute deviation.}
}
\value{
An S4 object of type \code{.pass.class} containing the data \code{Xi}, procedure parameter values and the results.
}
\description{
Implements the Proportion Adaptive Segment Selection procedure of REF HERE to (what does it do ???)
}
\examples{
library(anomaly)
data(simulated)
res<-pass(sim.data)
summary(res)
plot(res,variate_names=FALSE)

}
\references{
\insertRef{10.1093/biomet/ass059}{anomaly}
}
