% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pass.class.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,pass.class-method}
\title{Displays S4 objects produced by pass.}
\usage{
\S4method{show}{pass.class}(object)
}
\arguments{
\item{object}{An S4 object produced by \code{\link{pass}}.}
}
\description{
Displays a summary of the S4 object produced by \code{\link{capa}}, \code{\link{uvcapa}}, \code{\link{mvcapa}} or \code{\link{pass}}.
The information displayed is the same as that provided by the \code{summary} method. The method is used by the S4 system for automatic printing.
}
