% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_series.R, R/capa.R, R/pass.class.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,anomaly_series-method}
\alias{show,capa.class-method}
\alias{show,pass.class-method}
\title{Displays S4 objects produced by capa methods.}
\usage{
\S4method{show}{anomaly_series}(object)

\S4method{show}{capa.class}(object)

\S4method{show}{pass.class}(object)
}
\arguments{
\item{object}{An S4 class produced by \code{\link{capa}}, \code{\link{capa.uv}}, \code{\link{capa.mv}}, or \code{\link{pass}}.}
}
\description{
Displays S4 object produced by \code{\link{capa}}, \code{\link{capa.uv}}, \code{\link{capa.mv}}, and
\code{\link{pass}}. The information produced is the same as that provided by the summary method. The method is used by the S4 system for automatic printing.
}
\seealso{
\code{\link{capa}},\code{\link{capa.uv}},\code{\link{capa.mv}},\code{\link{pass}}.
}
