% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_series.R, R/capa.R, R/pass.class.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,anomaly_series-method}
\alias{summary,capa.class-method}
\alias{summary,capa.uv.class-method}
\alias{summary,capa.mv.class-method}
\alias{summary,pass.class-method}
\title{Summary of collective and point anomalies.}
\usage{
\S4method{summary}{anomaly_series}(object)

\S4method{summary}{capa.class}(object)

\S4method{summary}{capa.uv.class}(object)

\S4method{summary}{capa.mv.class}(object)

\S4method{summary}{pass.class}(object)
}
\arguments{
\item{object}{An S4 class produced by \code{\link{capa}}, \code{\link{capa.uv}}, \code{\link{capa.mv}}, or \code{\link{pass}}.}

\item{...}{Other parameters to be passed to summary methods.}
}
\description{
Summary methods for S4 objects returned by \code{\link{capa}}, \code{\link{capa.uv}}, \code{\link{capa.mv}},
and \code{\link{pass}}.  The output displayed depends on the type of S4 object passed to summary. For all types, the output indicates whether the data is univariate or
multivariate, the number of observations in the data, and the type of change being detected.


For an object produced by \code{\link{capa.uv}} or \code{\link{capa.mv}}, \code{summary} displays a summary of the analysis type used and the results
produced by \code{\link{point_anomalies}} and \code{\link{collective_anomalies}}.


For an object produced by \code{\link{capa}} is the same as for an object produced by \code{\link{capa.uv}}
or \code{\link{capa.mv}}.
}
\seealso{
\code{\link{capa}},\code{\link{capa.uv}},\code{\link{capa.mv}},,\code{\link{pass}},\code{\link{point_anomalies}},\code{\link{collective_anomalies}}.
}
